+++ cardioBackend/routers/readings.py
@@ -4,7 +4,7 @@ from fastapi.responses import Response, JSONResponse
 from sqlalchemy.orm import Session
 from cardioBackend import schemas
 from cardioBackend.oauth2 import get_current_user
-from cardioBackend.schemas import PyObjectId
+from cardioBackend.schemas import PyObjectId, Roles
 from cardioBackend.database import db
 from bson import ObjectId
 from typing import List
@@ -15,6 +15,8 @@ collection = db.readings
 
 @router.get("/", status_code=200)
 async def get_readings(current_user: schemas.User = Depends(get_current_user)):
+    if current_user.role != Roles.TECHNICIAN:
+        raise HTTPException(status_code=403, detail='Forbidden! Url is not permitted to this user.')
     arr = await collection.find().to_list(1000)
     readings = []
     for reading in arr:
@@ -23,6 +25,8 @@ async def get_readings(current_user: schemas.User = Depends(get_current_user)):
 
 @router.get("/patient/{id}", status_code=200)
 async def get_readings_by_patient_id(id, current_user: schemas.User = Depends(get_current_user)):
+    if current_user.role != Roles.TECHNICIAN or current_user.role != Roles.DOCTOR:
+        raise HTTPException(status_code=403, detail='Forbidden! Url is not permitted to this user.')
     arr = await collection.find({'patient_id': id}).to_list(1000)
     readings = []
     for reading in arr:
@@ -31,6 +35,8 @@ async def get_readings_by_patient_id(id, current_user: schemas.User = Depends(ge
 
 @router.get("/{id}", status_code=200)
 async def get_reading_by_id(id, current_user: schemas.User = Depends(get_current_user)):
+    if current_user.role != Roles.TECHNICIAN or current_user.role != Roles.DOCTOR:
+        raise HTTPException(status_code=403, detail='Forbidden! Url is not permitted to this user.')
     data = await collection.find_one({'_id': id})
     ecg_values = await db.ecgValues.find_one({'reading_id': id})
     if data is None:
@@ -39,6 +45,8 @@ async def get_reading_by_id(id, current_user: schemas.User = Depends(get_current
 
 @router.post("/", status_code=201)
 async def add_reading(request: schemas.Reading, current_user: schemas.User = Depends(get_current_user)):
+    if current_user.role != Roles.TECHNICIAN or current_user.role != Roles.DOCTOR:
+        raise HTTPException(status_code=403, detail='Forbidden! Url is not permitted to this user.')
     new_reading = schemas.ShowReading(
         patient_id =request.patient_id, 
         hospital_name =request.hospital_name, 
@@ -62,6 +70,8 @@ async def add_reading(request: schemas.Reading, current_user: schemas.User = Dep
 
 @router.put("/{id}", status_code=200)
 async def edit_reading_details(id, reading: schemas.ShowReading, current_user: schemas.User = Depends(get_current_user)):
+    if current_user.role != Roles.TECHNICIAN:
+        raise HTTPException(status_code=403, detail='Forbidden! Url is not permitted to this user.')
     reading = {k: v for k, v in reading.dict().items() if v is not None}
 
     if len(reading) >= 1:
@@ -80,6 +90,8 @@ async def edit_reading_details(id, reading: schemas.ShowReading, current_user: s
 
 @router.delete("/{id}", status_code=203)
 async def delete_reading(id, current_user: schemas.User = Depends(get_current_user)):
+    if current_user.role != Roles.TECHNICIAN:
+        raise HTTPException(status_code=403, detail='Forbidden! Url is not permitted to this user.')
     delete_result = await collection.delete_one({"_id":id})
     if delete_result.deleted_count == 1:
         return JSONResponse(status_code=status.HTTP_204_NO_CONTENT, content=[])
