+++ cardioBackend/routers/users.py
@@ -1,7 +1,9 @@
+import imp
 from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi.responses import Response, JSONResponse
 from sqlalchemy.orm import Session
 from cardioBackend import schemas
+from cardioBackend.schemas import Roles
 from cardioBackend.oauth2 import get_current_user
 from cardioBackend.database import db
 
@@ -10,6 +12,8 @@ collection = db.users
 
 @router.get("/", status_code=200)
 async def get_users(current_user: schemas.User = Depends(get_current_user)):
+    if current_user.role != Roles.ADMIN and current_user.role != Roles.TECHNICIAN:
+        raise HTTPException(status_code=403, detail='Forbidden! Url is not permitted to this user.')
     arr = await collection.find().to_list(1000)
     users = []
     for user in arr:
@@ -20,6 +24,13 @@ async def get_users(current_user: schemas.User = Depends(get_current_user)):
 async def get_me(current_user: schemas.User = Depends(get_current_user)):
     return current_user
 
+@router.get("/info", status_code=200)
+async def get_user_info(current_user: schemas.User = Depends(get_current_user)):
+    data = await collection.find_one({'_id': current_user.id})
+    if data is None:
+        raise HTTPException(status_code=404, detail='Oops! User not found.')
+    return schemas.User(**data)
+
 @router.get("/{id}", status_code=200)
 async def get_user_by_id(id, current_user: schemas.User = Depends(get_current_user)):
     data = await collection.find_one({'_id': id})
@@ -29,6 +40,8 @@ async def get_user_by_id(id, current_user: schemas.User = Depends(get_current_us
 
 @router.put("/{id}", status_code=200)
 async def edit_user_details(id, user: schemas.UpdateUser, current_user: schemas.User = Depends(get_current_user)):
+    if current_user.role != Roles.ADMIN or current_user.id != id:
+        raise HTTPException(status_code=403, detail='Forbidden! Url is not permitted to this user.')
     user = {k: v for k, v in user.dict().items() if v is not None}
 
     if len(user) >= 1:
@@ -47,6 +60,8 @@ async def edit_user_details(id, user: schemas.UpdateUser, current_user: schemas.
 
 @router.delete("/{id}", status_code=203)
 async def delete_user(id, current_user: schemas.User = Depends(get_current_user)):
+    if current_user.role != Roles.ADMIN or current_user.id != id:
+        raise HTTPException(status_code=403, detail='Forbidden! Url is not permitted to this user.')
     delete_result = await collection.delete_one({"_id":id})
     if delete_result.deleted_count == 1:
         return JSONResponse(status_code=status.HTTP_204_NO_CONTENT, content=[])
